+++ src/app/app.ts
@@ -64,7 +64,7 @@ function initializeClient(): Client {
 			Intents.FLAGS.DIRECT_MESSAGES,
 			Intents.FLAGS.DIRECT_MESSAGE_REACTIONS,
 		],
-		partials: ['MESSAGE', 'CHANNEL', 'REACTION'],
+		partials: ['MESSAGE', 'CHANNEL', 'REACTION', 'USER'],
 	};
 	return new Discord.Client(clientOptions);
 }