+++ CHANGELOG.md
@@ -2,7 +2,29 @@
 
 ## 1.10.0-SNAPSHOT
 
-1. Tweet spaces event upon startup
+1. Add banned names to username spam filter.
+2. DM flow triggered by GuildMemberAdd event
+   - CAPTCHA verification
+   - first quest messages + role assignments
+   - POAP distribution upon completion
+3. /first-quest config command introduced
+   - update FQ message content in database via DMs
+4. /first-quest poap-refill command introduced
+   - add option: add claim links to DB
+   - replace option: add claim links + retire existing claim links
+5. background services
+   - FQ rescue call service
+   - POAP claim link expiration reminder
+6. Fix node version heroku deployment
+7. Add /afk feature
+   - bot auto responds that mentioned users are AFK
+   - bot sends AFK users DM's with links to their messages
+8. Tweet spaces event upon startup
+
+## 1.9.3-RELEASE (2021-11-12)
+
+1. Add try/catch block to buffering
+   - display logdna error if missing token
 
 ## 1.9.2-RELEASE (2021-11-10)